state.jsとrouter.jsの中身のファイル分割
