+++ frontend/src/main.js
@@ -5,7 +5,7 @@ import ElementUI from 'element-ui'
 import locale from 'element-ui/lib/locale/lang/ja'
 import 'element-ui/lib/theme-chalk/index.css'
 import 'element-ui/lib/theme-chalk/display.css'
-import router from './router.js'
+import router from './router'
 import store from './store'
 
 Vue.config.productionTip = false
