+++ frontend/src/store/mutations.js
@@ -1,0 +1,5 @@
+export default {
+  todos(state, todos) {
+    state.todos = todos
+  }
+}
